Moved all schedule page to classadmin.
